requirements.txt
chromadb
ollama
python-frontmatter
pyyaml
